+++ debug-client/fabric-26.1.2/src/main/java/io/github/trethore/graphene/debug/GrapheneDebugMinecraftAccess.java
@@ -1,0 +1,18 @@
+package io.github.trethore.graphene.debug;
+
+import net.fabricmc.fabric.api.client.keymapping.v1.KeyMappingHelper;
+import net.minecraft.client.KeyMapping;
+import net.minecraft.client.Minecraft;
+import net.minecraft.client.gui.screens.Screen;
+
+final class GrapheneDebugMinecraftAccess {
+    private GrapheneDebugMinecraftAccess() {}
+
+    static void registerKeyMapping(KeyMapping mapping) {
+        KeyMappingHelper.registerKeyMapping(mapping);
+    }
+
+    static void showScreen(Minecraft minecraft, Screen screen) {
+        minecraft.setScreenAndShow(screen);
+    }
+}